requirements.txt
aiogram==3.4.1
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
aiohttp==3.9.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.1.0
yookassa==3.2.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
cryptography==42.0.2
jinja2==3.1.3